Implemented full create and delete post endpoint test
